projects/eval-TMC_DEMO: Updated spi driver for TMC dev drivers.
Added spi_master_readwrite_bytes that sends a single byte and
controls the chip select in manual way. If the last_transfer
parameter is set, then sets the chip select back to high,
otherwise keeps it low. The spi_master_config function spi_enable
parameter has to be set to 1 in order to this function to work
properly.

Signed-off-by: Villyam <Vilmoscsaba.Jozsa@analog.com>